pyproject.toml
[tool.poetry]
name = "typed_dotenv"
version = "1.0.1"
description = "Handle .env files with types"
authors = ["Ewen Le Bihan <hey@ewen.works>"]
repository = "https://github.com/ewen-lbh/python-typed-dotenv"
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8"
python-dotenv = "^0.14.0"
"ruamel.yaml" = { version = "^0.16.10", optional = true }
pydantic = { version = "^1.6.1", optional = true }
toml = {version = "^0.10.1", optional = true }
parse = "^1.16.0"

[tool.poetry.extras]
yaml = ["ruamel.yaml"]
pydantic = ["pydantic"]
toml = ["toml"]
all = ["ruamel.yaml", "pydantic", "toml"]

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.4"
pytest = "^8.1.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
